Oracle8iDialect.cs: Define atan2 and power to return double to harmonize with other dialects. Make the Linq provider also handle Math.Pow(). NH-3381.

+++ src/NHibernate/Dialect/Oracle8iDialect.cs
@@ -210,12 +210,12 @@ namespace NHibernate.Dialect
 			RegisterFunction("coalesce", new NvlFunction());
 
 			// Multi-param numeric dialect functions...
-			RegisterFunction("atan2", new StandardSQLFunction("atan2", NHibernateUtil.Single));
+			RegisterFunction("atan2", new StandardSQLFunction("atan2", NHibernateUtil.Double));
 			RegisterFunction("log", new StandardSQLFunction("log", NHibernateUtil.Int32));
 			RegisterFunction("mod", new StandardSQLFunction("mod", NHibernateUtil.Int32));
 			RegisterFunction("nvl", new StandardSQLFunction("nvl"));
 			RegisterFunction("nvl2", new StandardSQLFunction("nvl2"));
-			RegisterFunction("power", new StandardSQLFunction("power", NHibernateUtil.Single));
+			RegisterFunction("power", new StandardSQLFunction("power", NHibernateUtil.Double));
 
 			// Multi-param date dialect functions...
 			RegisterFunction("add_months", new StandardSQLFunction("add_months", NHibernateUtil.Date));
